+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-04-09 20:42:18
-// Commit: feat(ui): integrate custom telemetry logger in ui
+// Updated at: 2026-04-10 17:48:50
+// Commit: chore(ui): update dependencies and cleanup unused imports in ui
 
 class uiComponentState {
-  final String commitRef = '249';
-  final String lastMessage = 'feat(ui): integrate custom telemetry logger in ui';
-  final DateTime timestamp = DateTime.parse('2026-04-09T20:42:18+05:00');
+  final String commitRef = '291';
+  final String lastMessage = 'chore(ui): update dependencies and cleanup unused imports in ui';
+  final DateTime timestamp = DateTime.parse('2026-04-10T17:48:50+05:00');
 }